perf(tests): split the file-display icons off the file-type data
`file-display-helpers.ts` is mostly pure model-file-type data — eight server
modules import `primaryModelFileTypes` / `getPrimaryFileTypes` from it — but three
of its exports carry icon COMPONENTS, so those eight dragged
`@tabler/icons-react` in behind them.

`componentTypeConfig`, `fileFormatConfig` and `getFileIconConfig` move to
`file-display-icons.ts`; the four components that use them import from there.
Unit test files reaching @tabler: 135 -> 15.
